refactor: Select를 서버 컴포넌트로 변경 및 그에 따라 payements 페이지 라우팅 방식 개선
